% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process_funcs.R
\name{getCelllines}
\alias{getCelllines}
\title{Retrieve cell lines for corresponding cancer type}
\usage{
getCelllines(x, cellline_annot, meta_data)
}
\arguments{
\item{x}{primary site}

\item{cellline_annot}{cell line annotation data frame}

\item{meta_data}{information on different sub types for each primary site}
}
\value{
A vector of cell lines for each cancer type
}
\description{
The function takes in the annotation file, meta data and the cancer type and returns
a set of all cell lines for that cancer type.
}
