% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_test_funcs.R
\name{getPval}
\alias{getPval}
\title{P-value for a given pair of genes
Given the processed data object, a driver gene, and it's prospective partner, the
function returns the p-values for the wild type and mutated samples.}
\usage{
getPval(canc_data, driver_gene, sl_partner_gene)
}
\arguments{
\item{canc_data}{Processed data object for a given cancer type.}

\item{driver_gene}{The target mutated gene.}

\item{sl_partner_gene}{The corresponding SL partner}
}
\value{
a dataframe of driver gene with its corresponding SL partner, the p-value in WT samples,
p-value in mutated samples.
}
\description{
P-value for a given pair of genes
Given the processed data object, a driver gene, and it's prospective partner, the
function returns the p-values for the wild type and mutated samples.
}
