% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_test_funcs.R
\name{IH_CDF}
\alias{IH_CDF}
\title{CDF for Irwin-Hall distribution}
\usage{
IH_CDF(x, n)
}
\arguments{
\item{x}{normalised rank sum of viabilities in the mutated samples}

\item{n}{number of mutated samples}
}
\value{
the CDF value
}
\description{
The function computes the CDF for Irwin-Hall distribution which is the p-value for the
rank of viablities in the mutated samples.
}
