% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{plotSLBoxplot}
\alias{plotSLBoxplot}
\title{Function to plot annotated box-plots of SL pairs for a given cancer type}
\usage{
plotSLBoxplot(canc_data, hits_df, path_results, WT_pval_thresh = 0.2)
}
\arguments{
\item{canc_data}{Processed data object for a given cancer type}

\item{hits_df}{dataframe of SL pairs with their p-values returned by `identifySLHits` function}

\item{path_results}{The path to where the results should be stored}

\item{WT_pval_thresh}{Discard SL pairs with WT p-values less than this threshold. Default = 0.2}
}
\description{
Function to plot annotated box-plots of SL pairs for a given cancer type
}
