% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process_funcs.R
\name{prepareCNMat}
\alias{prepareCNMat}
\title{Prepare integer valued copy number matrix}
\usage{
prepareCNMat(CN_df, samples, driver_genes, x)
}
\arguments{
\item{CN_df}{log2 transformed copy number matrix}

\item{samples}{vector of cell lines of interest}

\item{driver_genes}{vector of driver or mutated genes of interest}

\item{x}{primary site. If `NULL` then ensure that the `samples` passed match the colnames of data}
}
\value{
Integer valued copy number matrix for selected cell lines and drivers
}
\description{
The function takes in log2 ratios of copy number as obtained from CCLE,
cell lines of interest and top drivers and returns an integer valued copy
number matrix
}
