% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process_funcs.R
\name{prepareCNMatGistic}
\alias{prepareCNMatGistic}
\title{Prepare integer valued copy number matrix from Gistic based copy number}
\usage{
prepareCNMatGistic(CN_df_gistic, celllines, top_drivers, x)
}
\arguments{
\item{CN_df_gistic}{Copy number in Gistic format}

\item{celllines}{vector of cell lines of interest}

\item{top_drivers}{vector of driver or mutated genes of interest}

\item{x}{primary site}
}
\value{
Integer valued copy number matrix for selected cell lines and drivers
}
\description{
The function takes in Gistic values of copy number as obtained from CCLE/TCGA,
cell lines of interest and top drivers and returns an integer valued copy
number matrix (0 = Homozygous deletion; 1 = Heterozygous deletion; 2 = Normal; etc.).
}
