% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process_funcs.R
\name{prepareDataObjects}
\alias{prepareDataObjects}
\title{Prepare the data object required for downstream analysis}
\usage{
prepareDataObjects(data, x, fdr = 0.05, min_Nmut = 2, all_cancers_mut_df,
  CN_mat, minNrcelllines = 5, cellline_annot, meta_data,
  essential_genes = NULL)
}
\arguments{
\item{data}{input data frame of cell line viabilities for different gene knockdowns}

\item{x}{primary site}

\item{fdr}{fdr cut-off for choosing the top drivers from mutSig2 list of drivers. Default = 0.05}

\item{min_Nmut}{lower bound of number of cell lines with mutations. Default = 2}

\item{all_cancers_mut_df}{MAF file from CCLE}

\item{CN_mat}{copy number matrix from CCLE}

\item{minNrcelllines}{lower bound of number of cell lines. Default = 5}

\item{cellline_annot}{cell line annotation data frame}

\item{meta_data}{information on different sub types for each primary site}

\item{essential_genes}{vector of essential genes}
}
\value{
An object for each cancer type
\describe{
   \item{viabilities}{matrix of viabilities for each cancer type}
   \item{mutations}{matrix of mutations in drivers for each cancer type}
   \item{mutation_annot}{annotations of the mutations}
   \item{primary_site}{cancer type}
}
}
\description{
The function processes the pan-cancer data and returns an object with viabilities matrix, mutation matrix,
mutation annotations and primary site for different types of cancers.
}
