% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process_funcs.R
\name{prepareMutMat}
\alias{prepareMutMat}
\title{Prepare mutation matrix}
\usage{
prepareMutMat(x, driver_genes, samples, all_cancers_mut_df)
}
\arguments{
\item{x}{primary site}

\item{driver_genes}{vector of interested driver genes}

\item{samples}{vector of cell lines or patient samples}

\item{all_cancers_mut_df}{dataframe of mutations with Hugo_symbol, Tumor_Sample_Barcode and type of mutation}
}
\value{
a mutation matrix for the given samples and genes
}
\description{
Function takes in a list of driver genes, samples and mutation/maf file
and returns a binary muration matrix
}
