% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_test_funcs.R
\name{ttestSLHits}
\alias{ttestSLHits}
\title{Retrieve list of mutation-specific synthetic lethal partners based on t test
for each type of cancer}
\usage{
ttestSLHits(canc_data)
}
\arguments{
\item{canc_data}{Processed data object for a given cancer type}
}
\value{
a dataframe of driver gene with its corresponding SL partner, the p-value in WT samples,
p-value in mutated samples, and the corresponding value after scaling.
}
\description{
Test based on t test to identify synthetic lethal (SL) partners
using viabilities.
}
