% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_test_funcs.R
\name{WilcoxSLHits}
\alias{WilcoxSLHits}
\title{Retrieve list of mutation-specific synthetic lethal partners based on Wilcoxon test
for each type of cancer}
\usage{
WilcoxSLHits(canc_data)
}
\arguments{
\item{canc_data}{Processed data object for a given cancer type}
}
\value{
a dataframe of driver gene with its corresponding SL partner, the p-value in WT samples,
p-value in mutated samples, and the corresponding value after scaling.
}
\description{
Test based on Wilcoxon test to identify synthetic lethal (SL) partners
in the mutated samples.
}
