% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process_funcs.R
\name{getEssentialGenes}
\alias{getEssentialGenes}
\title{Get a list of essential genes}
\usage{
getEssentialGenes(x, celllines, data, thresh = -3)
}
\arguments{
\item{x}{primary site. If `NULL` then ensure that the `celllines` passed match the colnames of data}

\item{celllines}{subset of interested celllines}

\item{data}{data frame of RSA normalised values}

\item{thresh}{threshold for defining essentiality. Default = -3}
}
\value{
a vector of essential genes
}
\description{
The function takes in RSA data and returns a vector of essential genes
}
