% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_test_funcs.R
\name{identifySLHits}
\alias{identifySLHits}
\title{Retrieve list of mutation-specific synthetic lethal partners for each type of cancer
Rank test based on Irwin-Hall distribution to identify synthetic lethal (SL) partners
in the mutated samples. The function also performs a two sided wilcoxon or t-test
on the wild type samples to give a p-value required for filtering.}
\usage{
identifySLHits(canc_data, n_cand = 100, qval_thresh = 1, path_results,
  WT_pval_thresh)
}
\arguments{
\item{canc_data}{Processed data object for a given cancer type}

\item{n_cand}{first 'n' number of SL partners to be tested after ranking. Default = 100
This is just to save on computation time since after multiple testing correction we will be
filtering a lot of them out.}

\item{qval_thresh}{The number of false positives allowed after correction. Default = 1}

\item{path_results}{The path to where the results should be stored}

\item{WT_pval_thresh}{Discard SL pairs with WT p-values less than this threshold. Default = 0.2
Required by `plotSLBoxplot` function}
}
\value{
a dataframe of driver gene with its corresponding SL partner, the p-value in WT samples,
p-value in mutated samples and the corresponding value after correction.
}
\description{
Retrieve list of mutation-specific synthetic lethal partners for each type of cancer
Rank test based on Irwin-Hall distribution to identify synthetic lethal (SL) partners
in the mutated samples. The function also performs a two sided wilcoxon or t-test
on the wild type samples to give a p-value required for filtering.
}
