% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process_funcs.R
\name{prepareCNMat}
\alias{prepareCNMat}
\title{Prepare integer valued copy number matrix}
\usage{
prepareCNMat(CN_mat, celllines, top_drivers, x)
}
\arguments{
\item{CN_mat}{log2 transformed copy number matrix}

\item{celllines}{vector of cell lines of interest}

\item{top_drivers}{vector of driver or mutated genes of interest}

\item{x}{primary site}
}
\value{
Integer valued copy number matrix for selected cell lines and drivers
}
\description{
The function takes in log2 ratios of copy number as obtained from CCLE,
cell lines of interest and top drivers and returns an integer valued copy 
number matrix
}
