% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process_funcs.R
\name{prepareDataObjects}
\alias{prepareDataObjects}
\title{Prepare the data object required for downstream analysis}
\usage{
prepareDataObjects(data, x, fdr = 0.05, min_Nmut = 2, all_cancers_mut_df,
  CN_df, gistic = FALSE, CN_Thr = 2, minNrcelllines = 5, celllines,
  meta_data, essential_genes = NULL)
}
\arguments{
\item{data}{input data frame of cell line viabilities for different gene knockdowns}

\item{x}{primary site}

\item{fdr}{fdr cut-off for choosing the top drivers from mutSig2 list of drivers. Default = 0.05}

\item{min_Nmut}{lower bound of number of cell lines with mutations. Default = 2}

\item{all_cancers_mut_df}{MAF file from CCLE}

\item{CN_df}{copy number dataframe from CCLE}

\item{gistic}{Logical variable checking if copy number is based on Gistic. Default = FALSE}

\item{CN_Thr}{threshold for using CN data. Values: 0 = Homozygous and heterozygous deletions ; 1 = Homozygous deletions only; 2 = No copy number used (default)}

\item{minNrcelllines}{lower bound of number of cell lines. Default = 5}

\item{celllines}{vector of interested celllines}

\item{meta_data}{information on different sub types for each primary site}

\item{essential_genes}{vector of essential genes}
}
\value{
An object for each cancer type
\describe{
   \item{viabilities}{dataframe of viabilities for each cancer type}
   \item{mutations}{matrix of mutations in drivers for each cancer type}
   \item{CNalterations}{matrix of non-negative copy number alterations of drivers for each cancer type}
   \item{mutation_annot}{annotations of the mutations}
   \item{primary_site}{cancer type}
}
}
\description{
The function processes the pan-cancer data and returns an object with viabilities matrix, mutation matrix,
mutation annotations and primary site for different types of cancers.
}
